Server/Feature: Added on create comment server-side  data validation

+++ servers/backend-server/src/repositories/comment-repositories.js
@@ -1,12 +1,28 @@
 import Comment from "../models/Comment.js";
+import User from "../models/User.js";
+import Post from "../models/Post.js";
 
 import { postedOnDateConverter } from "../utils/date-time-utils.js";
+import mongoose from "mongoose";
 
 const COMMONLY_NEEDED_PARAMS = 'firstName lastName imageUrl'
 
 async function create(commentData) {
     // The line below offsets the time with 2+ hours as new Date() is 2 hours behind;
     const creationTime = new Date(Date.now() - new Date().getTimezoneOffset() * 60000);
+
+    const [ownerId, parentPostId] = await Promise.all([
+        User.findById(commentData.owner).select('_id').lean(),
+        Post.findById(commentData.onPost).select('_id').lean(),
+    ])
+
+    if (!ownerId) {
+        throw new Error("Could not create comment: Creator id is invalid/missing");
+    } else if (!parentPostId) {
+        throw new Error("Could not create comment: Parent post id is invalid/missing");
+    } else if (!commentData.text) {
+        return;
+    }
 
     commentData.createdAt = creationTime;
     commentData.postedOn = postedOnDateConverter(creationTime);
@@ -18,7 +34,7 @@ async function create(commentData) {
         select: COMMONLY_NEEDED_PARAMS
     });
 
-    return newComment;
+    return { newComment };
 }
 
 async function removeSpecific(commentId) {
@@ -28,6 +44,10 @@ async function removeSpecific(commentId) {
 }
 
 async function removeAllSharingPost(postId) {
+    if (!postId) {
+        throw new Error("Parent post id is missing!");
+    }
+
     //Pretty scary stuff over here, deleting so many documents
     await Comment.deleteMany({ onPost: postId });
 }
